requirements/win.txt
alembic==0.8.8
amqp==1.4.9
anyjson==0.3.3
arrow==0.8.0
Babel==2.3.4
billiard==3.3.0.23
blinker==1.4
celery==3.1.23
cffi==1.9.1
click==6.6
cryptography==1.6
cssselect==1.0.0
cssutils==1.0.1
Cython==0.25.1
decorator==4.0.10
enum34==1.1.6
-e git+https://github.com/unifispot/facebook-sdk@ad0a36300bc02720d052072e1d2679a4716003f4#egg=facebook_sdk
factory-boy==2.7.0
fake-factory==0.7.2
Flask==0.11.1
Flask-Assets==0.12
Flask-Babel==0.11.1
Flask-BabelEx==0.9.3
Flask-Classful==0.13.1
Flask-Classy==0.6.10
Flask-Login==0.3.2
Flask-Mail==0.9.1
-e git+https://github.com/unifispot/unifispot-menu#egg=unifispot-menu
Flask-Principal==0.4.0
Flask-QRcode==0.8.0
Flask-Redis==0.3.0
Flask-Script==2.0.5
Flask-Security==1.7.5
Flask-SQLAlchemy==2.1
Flask-SSLify==0.1.5
Flask-Themes2==0.1.4
Flask-WTF==0.13.1
geoip2==2.4.0
idna==2.1
ipaddress==1.0.17
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.37
lxml==3.6.4
Mako==1.0.4
MarkupSafe==0.23
maxminddb==1.2.1
passlib==1.6.5
paypalrestsdk==1.12.0
Pillow==3.4.2
premailer==3.0.1
py==1.4.31
pyasn1==0.1.9
-e git+https://github.com/unifispot/pycard@7847297287d847ec4f21ef968e9c654017f9d532#egg=pycard
pycparser==2.17
pyOpenSSL==16.2.0
pytest==3.0.3
pymongo==3.4.0
pytest-flask==0.10.0
python-dateutil==2.5.3
python-editor==1.0.1
pytz==2016.7
qrcode==5.3
redis==2.10.5
requests==2.11.1
requests-cache==0.4.12
simplejson==3.8.2
six==1.10.0
speaklater==1.3
SQLAlchemy==1.1.2
stripe==1.42.0
timezones==1.9.6
-e git+https://github.com/unifispot/unifispot-datatables.git@2c82cb81aae1a55bc2ce0cf8d7942e58f5a86d81#egg=unifispot_datatables
-e git+https://github.com/unifispot/unifispot-menu.git@a475af75e825fdf968dcaefb0f2284169f647580#egg=Unifispot_Menu
-e git+https://github.com/unifispot/unifispot-migrate.git@e99a170c7c49c2d02ec4540521fef4d443028221#egg=Unifispot_Migrate
validators==0.11.0
webassets==0.12.0
Werkzeug==0.11.11
WTForms==2.1
